+++ CHANGELOG.md
@@ -3,10 +3,15 @@
 - skater-name correction
 - %3.2f, %d for showing table data / centering
 - content_for on <title>
+- total base_value in score
 - scores.total_bv, total_goe
-- elements/show.csv should include scores key
+  - score_parser recognize total base value
+- score controler: competitions.season to check
+- spec:
+  - all filters to work
 
 ## 1.0.0-pre3
+- elements/show.csv should include scores key
 
 ## 1.0.0-pre2